+++ R/RunMistyR.R
@@ -39,6 +39,7 @@
 #'
 #' @return A `Seurat` object with results stored in `srt@tools[[tool_name]]`
 #' when `store_results = TRUE`.
+#' @concept spatial-producer
 #' @export
 #'
 #' @examples