+++ R/standard_scop.R
@@ -113,6 +113,7 @@
 #'
 #' @return A `Seurat` object.
 #'
+#' @concept spatial-producer
 #' @export
 #'
 #' @examples